Convert to Typescript
Move handlers to separate folder.
Various refactorings.
