+++ .gitignore
@@ -8,3 +8,6 @@ dist/
 .DS_Store
 
 src/data/strava_config
+
+*/**/*.js
+*/**/*.js.map
