+++ gulpfile.js
@@ -1,6 +1,9 @@
-var gulp   = require('gulp');
+var gulp = require('gulp');
 var lambda = require('gulp-awslambda');
-var zip    = require('gulp-zip');
+var merge = require('merge2');
+var ts = require("gulp-typescript");
+var tsProject = ts.createProject("tsconfig.json");
+var zip = require('gulp-zip');
 
 
 /**
@@ -9,7 +12,7 @@ var zip    = require('gulp-zip');
  * require the name of the function (see task below).
  */
 var lambda_params = {
-    FunctionName: 'mountaintop',
+    FunctionName: 'mountain-top-0-2',
     Role: '[YOUR LAMBDA EXEC ROLE HERE]'
 };
 
@@ -18,8 +21,10 @@ var opts = {
 };
 
 gulp.task('lambda', function() {
-    return gulp.src(['src/**'])
-        .pipe(zip('archive.zip'))
+    return merge(
+        tsProject.src().pipe(tsProject()),
+        gulp.src(['src/node_modules/**'], {base: 'src'})
+    ).pipe(zip('archive.zip'))
         //.pipe(lambda(lambda_params, opts))
         .pipe(lambda(lambda_params.FunctionName, opts))
         .pipe(gulp.dest('dist/'));
