+++ package.json
@@ -1,14 +1,21 @@
 {
   "name": "mountain-top",
   "version": "0.1.3",
-  "license": "LICENSE.md",
+  "license": "MIT",
   "readme": "README.md",
+  "repository": {
+    "type": "git",
+    "url": "git://github.com/nickcoury/mountain-top.git"
+  },
   "scripts": {
     "preinstall": "cd src && npm i && cd .."
   },
   "devDependencies": {
     "gulp": "^3.9.1",
     "gulp-awslambda": "^0.4.0",
-    "gulp-zip": "^3.2.0"
+    "gulp-typescript": "^3.1.3",
+    "gulp-zip": "^3.2.0",
+    "merge2": "^1.0.2",
+    "typescript": "^2.1.4"
   }
 }
