+++ src/types.ts
@@ -1,0 +1,10 @@
+interface Time {
+    seconds: number;
+    minutes: number;
+    hours: number;
+    days: number;
+}
+
+export {
+    Time
+}
