--- util/schema.js
@@ -1,3 +1,0 @@
-var lambda = require('../src/index.js');
-
-console.log(lambda.alexa.schema());
