+++ util/utterances.ts
@@ -1,0 +1,5 @@
+(function() {
+    let lambda = require('../src/index.js')
+        ;
+    console.log(lambda.alexa.utterances());
+})();
